pyproject.toml
[project]
name = "fractal-drawing"
version = "0.0.0"
description = "framework for fractal drawing "
authors = [
    { name = "Elin Huliiev" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "pandas-stubs>=2.2.0",
    "pre-commit==4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fractaldrawing"]
